ucspi-tcp6-1.02

+++ buffer_get.c
@@ -2,7 +2,7 @@
 #include "byte.h"
 #include "error.h"
 
-static int oneread(int (*op)(),int fd,char *buf,unsigned int len)
+static ssize_t oneread(ssize_t (*op)(),int fd,char *buf,unsigned int len)
 {
   int r;
